requirements.txt
fabric2
etcd3
